Corrected and reorganized documentation for SqlServerDatabaseMail
DisplayName is used for the display name of the email address, for example, the "SQL Server" part of a `From: "SQL Server" <mssql@example.com>` header. This was incorrectly indicated to be a display name for the mail server FQDN.

Also clarified some description and name fields. Tried to catch this in the psm1 file as well as the readme.

I also reordered the parameters listed in the readme for this resource since they seemed a bit random before.