Updated Get-ServiceObject to find SSIS service
Get-ServiceObject calls the Get-SqlServiceName function which searches the registry for the specified service type.  For SSIS, it correctly finds the service and returns MsDtsServer.  Get-ServiceObject then searches the $managedComputer.Services collection comparing the Name property with the returned value and fails as the collection contains MsDtsServer130.  Changed the operator from -eq to -like with an * after $serviceNameFilter.  This should still function correctly with instances as well.